+++ src/main/java/br/com/anteros/automacao/bancaria/campolivre/CLBancoDoBrasilNN17Convenio6.java
@@ -100,7 +100,7 @@ class CLBancoDoBrasilNN17Convenio6 extends AbstractCLBancoDoBrasil{
 		ContaBancaria conta = titulo.getContaBancaria();
 		String nossoNumero = titulo.getNossoNumero();
 		
-		this.add(new FixedField<Integer>(conta.getNumeroDaConta().getCodigoDaConta(), 6, Fillers.ZERO_LEFT));
+		this.add(new FixedField<Long>(conta.getNumeroDaConta().getCodigoDaConta(), 6, Fillers.ZERO_LEFT));
 		
 		this.add(new FixedField<String>(nossoNumero, 17, Fillers.ZERO_LEFT));
 		this.add(new FixedField<Integer>(SERVICO, 2));
