+++ src/main/java/br/com/anteros/automacao/bancaria/campolivre/CLBanrisulCobrancaNaoRegistrada.java
@@ -121,7 +121,7 @@ class CLBanrisulCobrancaNaoRegistrada extends AbstractCLBanrisul {
 		this.add(new FixedField<Integer>(2, 1));
 		this.add(new FixedField<String>("1", 1));
 		this.add(new FixedField<Integer>(titulo.getContaBancaria().getAgencia().getCodigo(), 4, Fillers.ZERO_LEFT));
-		this.add(new FixedField<Integer>(titulo.getContaBancaria().getNumeroDaConta().getCodigoDaConta(), 7, Fillers.ZERO_LEFT));
+		this.add(new FixedField<Long>(titulo.getContaBancaria().getNumeroDaConta().getCodigoDaConta(), 7, Fillers.ZERO_LEFT));
 		this.add(new FixedField<Integer>(Integer.valueOf(titulo.getNossoNumero()), 8, Fillers.ZERO_LEFT));
 		this.add(new FixedField<String>("40", 2));
