+++ src/main/java/br/com/anteros/automacao/bancaria/campolivre/CLCaixaEconomicaFederalSINCO.java
@@ -91,7 +91,7 @@ class CLCaixaEconomicaFederalSINCO extends AbstractCLCaixaEconomicaFederal {
 		
 		this.add(new FixedField<Integer>(1, 1));
 		
-		this.add(new FixedField<Integer>(conta.getNumeroDaConta().getCodigoDaConta(), 6, Fillers.ZERO_LEFT));
+		this.add(new FixedField<Long>(conta.getNumeroDaConta().getCodigoDaConta(), 6, Fillers.ZERO_LEFT));
 		
 		this.add(new FixedField<Integer>(9, 1));
 		this.add(new FixedField<String>(nossoNumero, 17));
