+++ src/main/java/br/com/anteros/integracao/bancaria/boleto/campo/BoletoInfoViewSicredi.java
@@ -102,7 +102,7 @@ public class BoletoInfoViewSicredi extends AbstractBoletoInfoCampoView{
 	private String getAgenciaCodigoCedente(){
 		Integer agencia = super.getBoleto().getTitulo().getContaBancaria().getAgencia().getCodigo();
 		Integer posto = super.getBoleto().getTitulo().getParametrosBancarios().getValor(ParametroBancoSicredi.POSTO_DA_AGENCIA);
-		Integer codigoCedente = super.getBoleto().getTitulo().getContaBancaria().getNumeroDaConta().getCodigoDaConta();
+		Long codigoCedente = super.getBoleto().getTitulo().getContaBancaria().getNumeroDaConta().getCodigoDaConta();
 		
 		return String.format("%04d.%02d.%05d", agencia, posto, codigoCedente);
 	}
